added shoes app and moved models out of api to shoes app
