+++ pyproject.toml
@@ -7,6 +7,7 @@ authors = ["Justin Miller <indymiller@gmail.com>"]
 [tool.poetry.dependencies]
 python = "^3.8"
 django = "^3.0.7"
+djangorestframework = "^3.11.0"
 
 [tool.poetry.dev-dependencies]
